US128094 Add start/end date observers
As part of adding the new W2D component to Brightspace for Parents, we
need to be able to change the date range W2D is showing, as this was a
feature of the old upcoming work widget in BfP. W2D mostly works in this
regard, but it doesn't fetch the new date range when an updated
start/end date is set on it. This change resolves that by re-fetching
whenever these parameters change.

This also includes a minor refactor by moving some common behaviour into
`_fetchRoutedState`, as it was needed by two existing functions, as well
as the two new ones.