US128094 Review changes
Might as well factor out the condition here as well - `setSirenEntity`
was using a slightly different one, but a) it's setting `this._href` so
that should be good, and b) `this._token` being falsy would fail anyway,
so might as well use the same check.

This also makes the startDate observer end-date-aware, and vice versa -
this is to avoid us making a call when only one of the two has been set,
as this would result in a 400.